+++ lib/SPVM/Sys/Time/Timespec.c
@@ -48,7 +48,7 @@ int32_t SPVM__Sys__Time__Timespec__set_tv_sec(SPVM_ENV* env, SPVM_VALUE* stack) 
   
   void* obj_tv = stack[0].oval;
   
-  int32_t tv_sec = stack[1].lval;
+  int64_t tv_sec = stack[1].lval;
   
   struct timespec* st_tv = env->get_pointer(env, stack, obj_tv);
   
@@ -72,7 +72,7 @@ int32_t SPVM__Sys__Time__Timespec__set_tv_nsec(SPVM_ENV* env, SPVM_VALUE* stack)
   
   void* obj_tv = stack[0].oval;
   
-  int32_t tv_nsec = stack[1].lval;
+  int64_t tv_nsec = stack[1].lval;
   
   struct timespec* st_tv = env->get_pointer(env, stack, obj_tv);
